Backend/requirements.txt
fastapi[all]
uvicorn
pydantic
python-dotenv
sqlalchemy
bcrypt
psycopg2-binary
pyjwt